[NWPS-1581] Link validator refactoring
