+++ repository-data/application/src/main/resources/hcm-config/configuration/modules/validation.yaml
@@ -51,6 +51,3 @@ definitions:
     /hippo:configuration/hippo:modules/validation/hippo:moduleconfig/hee:mandatory-training-journey-validator:
       jcr:primaryType: hipposys:moduleconfig
       hipposys:className: uk.nhs.hee.web.validation.validator.MandatoryTrainingJourneyValidator
-    /hippo:configuration/hippo:modules/validation/hippo:moduleconfig/hee:links-validator:
-      jcr:primaryType: hipposys:moduleconfig
-      hipposys:className: uk.nhs.hee.web.validation.validator.LinksValidator
