+++ repository-data/application/src/main/resources/hcm-config/configuration/translations/cms.yaml
@@ -47,7 +47,6 @@ definitions:
       hee:mandatory-training-journey-validator: When 'Training journey summary' field
         is filled in, you must select at least one link from either 'Prerequisite'
         or 'Optional Route'
-      hee:links-validator: Please enter either Link Document or URL.
     /hippo:configuration/hippo:translations/hippo:cms/advanced-search-filters:
       jcr:primaryType: hipposys:resourcebundles
       /en:
